Quickfix (#11)
* Change dtype from float to torch.float32

* Simplify calculation of final acc and test_loss

* Change dtype from torch.float32 to simply flaot

* Add Depatchify class to reassamble image from patches

* Test Patchify via Depatchify. Images are not displayed by default

+++ tests/test_model_components.py
@@ -9,7 +9,7 @@ def synthetic_data():
     batch_size = 16
     emb_dim = 64
     context_length = 10
-    data = torch.rand(batch_size, context_length, emb_dim, dtype=torch.float)
+    data = torch.rand(batch_size, context_length, emb_dim)
     return data
 
 @pytest.fixture
